requirements.txt
# GUI Framework
PyQt6==6.9.1

# Database
psycopg
sqlalchemy==2.0.25

# Environment Management
python-dotenv==1.0.0

# Data Validation
pydantic

# Utilities
pandas
openpyxl==3.1.2  # For Excel export
reportlab==4.0.8  # For PDF reports

# Development Tools (optional)
pytest==7.4.3
pytest-qt==4.2.0
black==23.12.1
flake8==6.1.0

# Logging
structlog==23.2.0